Remove inCheck.cpp, create move.cpp (#67)
* correct ranks naming, code cleanup, no functional change

* transfer move structs to move.cpp

* transfer make move into board.cpp

* add moveGen related functions from inCheck there

* finally remove inCheck

* update everything else to be in line with recent refactors
