+++ tests/CMakeLists.txt
@@ -22,12 +22,11 @@ add_executable(
     allTests
     testBitBoard.cpp
     testBoard.cpp
-    testInCheck.cpp
     testMoveGen.cpp
     testSearch.cpp
     ../src/bitboard.cpp
+    ../src/move.cpp
     ../src/board.cpp
-    ../src/inCheck.cpp
     ../src/moveGen.cpp
     ../src/search.cpp
     ../src/eval.cpp
